+++ CHANGELOG.md
@@ -32,9 +32,9 @@
 - Bump Android Gradle Plugin from v6.3.0 to v6.5.0 ([#1338](https://github.com/getsentry/sentry-unreal/pull/1338), [#1355](https://github.com/getsentry/sentry-unreal/pull/1355))
   - [changelog](https://github.com/getsentry/sentry-android-gradle-plugin/blob/main/CHANGELOG.md#650)
   - [diff](https://github.com/getsentry/sentry-android-gradle-plugin/compare/6.3.0...6.5.0)
-- Bump Crash Reporter from v0.1.0 to v0.1.1 ([#1344](https://github.com/getsentry/sentry-unreal/pull/1344))
-  - [changelog](https://github.com/getsentry/sentry-desktop-crash-reporter/blob/main/CHANGELOG.md#011)
-  - [diff](https://github.com/getsentry/sentry-desktop-crash-reporter/compare/0.1.0...0.1.1)
+- Bump Crash Reporter from v0.1.0 to v0.2.0 ([#1344](https://github.com/getsentry/sentry-unreal/pull/1344), [#1359](https://github.com/getsentry/sentry-unreal/pull/1359))
+  - [changelog](https://github.com/getsentry/sentry-desktop-crash-reporter/blob/main/CHANGELOG.md#020)
+  - [diff](https://github.com/getsentry/sentry-desktop-crash-reporter/compare/0.1.0...0.2.0)
 
 ## 1.10.0